requirements.txt
streamlit
tensorflow
tensorflow-hub
opencv-python-headless
numpy
Pillow
matplotlib